requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
cachetools==5.5.0
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
ecdsa==0.19.0
pyasn1==0.6.1
python-decouple==3.8
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
